Introduced persistent settings for language and backup language Re-written status management code for search and bookmark fragments as well as bookmark import activity
There is a race condition affecting search results order when switching back to English from another language, which will have to be fixed in a further commit.

Some flickering is also present because the backup language is not present in the data returned from the database but passed as a parameter to the list adapter, requiring manual refreshes.